refactor migration functions
